Fix T200 default values
t200_ms is an unsigned int [7] array, while the oml_default_t200_ms was
an uint8_t[7] array, which we memcpy() to the former as default
initializer.  Fix this by turning oml_default_t200_ms into unsigned int,
too.

+++ include/osmo-bts/oml.h
@@ -40,6 +40,6 @@ int oml_mo_fom_ack_nack(struct gsm_abis_mo *mo, uint8_t orig_msg_type,
 
 /* Configure LAPDm T200 timers for this lchan according to OML */
 int oml_set_lchan_t200(struct gsm_lchan *lchan);
-extern const uint8_t oml_default_t200_ms[7];
+extern const unsigned int oml_default_t200_ms[7];
 
 #endif // _OML_H */
